initial file cleanup
